8.11.2022 17:42 //FCariKarta AltHesap alanı eklendi.

+++ MEYPAK.Entity/PocoModels/CARI/PocoCARIALTHES.cs
@@ -8,10 +8,12 @@ namespace MEYPAK.Entity.PocoModels.CARI
 { 
     public class PocoCARIALTHES:SUPERPOCOMODEL
     {
+
+      
         public string adi { get; set; }
         public string kod { get; set; }
         public int dovizid { get; set; }
-        public byte aktif { get; set; }
+        public int aktif { get; set; }
 
     }
 }
